pyproject.toml
[project]
name = "file-upload-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "pydantic>=2.11.3",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]
